requirements.txt
openai
pyyaml
python-docx
cmakeast
pyinstaller